Move CaptureLocationTaskResult
