+++ ground/src/test/java/com/google/android/ground/ui/datacollection/tasks/location/CaptureLocationTaskFragmentTest.kt
@@ -26,7 +26,6 @@ import com.google.android.ground.R
 import com.google.android.ground.model.geometry.Coordinates
 import com.google.android.ground.model.geometry.Point
 import com.google.android.ground.model.job.Job
-import com.google.android.ground.model.submission.CaptureLocationTaskResult
 import com.google.android.ground.model.task.Task
 import com.google.android.ground.ui.common.ViewModelFactory
 import com.google.android.ground.ui.datacollection.DataCollectionViewModel
